move PushManager to separate file
