+++ deltachat-ios/Calls/CallManager.swift
@@ -1,39 +1,13 @@
 import CallKit
-import PushKit
 import DcCore
-
-class PushManager: NSObject, PKPushRegistryDelegate {
-    var pushRegistry: PKPushRegistry?
-
-    func registerForVoIPPushes() {
-        // registering for VoIP pushes is needed to enable the didReceiveIncomingPushWith callback,
-        // which is called via reportNewIncomingVoIPPushPayload from the regular NSE
-        pushRegistry = PKPushRegistry(queue: DispatchQueue.main)
-        pushRegistry?.delegate = self
-        pushRegistry?.desiredPushTypes = [.voIP]
-    }
-
-    func pushRegistry(_ registry: PKPushRegistry, didUpdate pushCredentials: PKPushCredentials, for type: PKPushType) {
-        // VoIP tokens are not used
-        logger.info("VoIP token received")
-    }
-
-    func pushRegistry(_ registry: PKPushRegistry, didReceiveIncomingPushWith payload: PKPushPayload, for type: PKPushType) {
-        let callInfo = payload.dictionaryPayload
-        logger.info("didReceiveIncomingPushWith: \(callInfo)")
-        guard let accountId = callInfo["account_id"] as? Int,
-              let msgId = callInfo["message_id"] as? Int else { return }
-        CallManager.shared.reportIncomingCall(accountId: accountId, msgId: msgId)
-    }
-}
 
 class CallManager: NSObject, CXProviderDelegate {
     static let shared = CallManager()
     private let provider: CXProvider
-    private let pushManager: PushManager
+    private let pushManager: VoIPPushManager
 
     override init() {
-        pushManager = PushManager()
+        pushManager = VoIPPushManager()
         pushManager.registerForVoIPPushes()
 
         let configuration = CXProviderConfiguration()
